Empty-board cycle recorder slice 2b: subset fan-out
Adds full subset-fan-out branching at the spec's two terminal cases:
  - turn 5 (no legal plays exist): pass + every distinct exchange subset
  - turn 6 always: pass + every distinct exchange subset + best-play
Turns 3, 4 still K=3 conditional on is_pass; turn 5 with plays still K=3.

Branch_id widened to uint64 with 8 bits per fork (was 4) so up to ~129
slots fit per fork. Slot 0 always pass; slots 1..N enumerated actions
in equity-sorted movegen order; subset-fan-out turn-6 best-play in
slot N+1.

Three correctness fixes uncovered while validating turn-6 fan-out:
  - eb_action_buf is per-runner, but inner forks during recursion
    overwrote it. Snapshot to stack-local Move array before iterating.
  - Per-player move_lists are sized to num_plays (often 1), which prunes
    MOVE_RECORD_ALL to top-1. Added a dedicated worker-level eb_move_list
    sized to 2000.
  - generate_moves with MOVE_RECORD_ALL leaves the list in heap order
    (min-heap by equity); call move_list_sort_moves so slot 1 is the
    top-equity action of its category, not the worst.

Verified on a 100-pair smoke run: pair 0 reached turn 6 with full
129-action fan-out (pass + 127 distinct exchange subsets + best-play).

+++ src/ent/empty_board.c
@@ -47,14 +47,14 @@ void empty_board_recorder_destroy(EmptyBoardRecorder *r) {
 }
 
 void empty_board_recorder_write(
-    EmptyBoardRecorder *r, uint64_t pair_id, int branch_id,
+    EmptyBoardRecorder *r, uint64_t pair_id, uint64_t branch_id,
     int turn_on_empty_board, const char *rack, const uint8_t bag_counts[27],
     const char *opp_action_history, int action_kind, const char *action_repr,
     int action_size, int eventual_outcome, int eventual_margin) {
   if (!r) return;
   pthread_mutex_lock(&r->mutex);
-  fprintf(r->out_file, "%llu,%d,%d,%s", (unsigned long long)pair_id, branch_id,
-          turn_on_empty_board, rack ? rack : "");
+  fprintf(r->out_file, "%llu,%llu,%d,%s", (unsigned long long)pair_id,
+          (unsigned long long)branch_id, turn_on_empty_board, rack ? rack : "");
   for (int i = 0; i < EB_TILE_TYPES; i++) {
     fprintf(r->out_file, ",%u", (unsigned)bag_counts[i]);
   }
